Updates api baseurl, formatting, consolidate article-related functions in one module
