+++ .env
@@ -1,1 +1,1 @@
-REACT_APP_BASEURL = "https://reactive-herald-api.herokuapp.com/api/v1"
+REACT_APP_BASEURL = "https://reactive-herald-api.herokuapp.com/api"
