+++ .env.development
@@ -1,1 +1,1 @@
-REACT_APP_BASEURL = "http://localhost:3000/api/v1"
+REACT_APP_BASEURL = "http://localhost:3000/api"
